Renamed lwutil to lwdebug
